backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
xgboost
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
joblib
